Splunk_TA_microsoft-cloudservices/lib/requirements.txt
adal==1.2.7 ; python_version >= "3.7" and python_version < "4.0"
azure-common==1.1.27 ; python_version >= "3.7" and python_version < "4.0"
azure-core==1.23.1 ; python_version >= "3.7" and python_version < "4.0"
azure-data-tables==12.4.0 ; python_version >= "3.7" and python_version < "4.0"
azure-eventhub-checkpointstoreblob==1.1.4 ; python_version >= "3.7" and python_version < "4.0"
azure-eventhub==5.10.1 ; python_version >= "3.7" and python_version < "4.0"
azure-identity==1.10.0 ; python_version >= "3.7" and python_version < "4.0"
azure-mgmt-core==1.3.1 ; python_version >= "3.7" and python_version < "4.0"
azure-mgmt-resource==21.1.0 ; python_version >= "3.7" and python_version < "4.0"
azure-storage-blob==12.12.0 ; python_version >= "3.7" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.7" and python_version < "4.0"
cffi==1.15.0 ; python_version >= "3.7" and python_version < "4.0"
charset-normalizer==2.0.10 ; python_version >= "3.7" and python_version < "4"
configparser==5.2.0 ; python_version >= "3.7" and python_version < "4.0"
cryptography==3.3.2 ; python_version >= "3.7" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.7" and python_version < "4.0"
future==0.18.3 ; python_version >= "3.7" and python_version < "4.0"
httplib2==0.20.4 ; python_version >= "3.7" and python_version < "4.0"
idna==3.3 ; python_version >= "3.7" and python_version < "4"
iso8601==1.0.2 ; python_version >= "3.7" and python_version < "4.0"
isodate==0.6.1 ; python_version >= "3.7" and python_version < "4.0"
msal-extensions==1.0.0 ; python_version >= "3.7" and python_version < "4.0"
msal==1.16.0 ; python_version >= "3.7" and python_version < "4.0"
msrest==0.6.21 ; python_version >= "3.7" and python_version < "4.0"
msrestazure==0.6.4 ; python_version >= "3.7" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.7" and python_version < "4.0"
portalocker==1.7.1 ; python_version >= "3.7" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.7" and python_version < "4.0"
pyjwt==2.4.0 ; python_version >= "3.7" and python_version < "4.0"
pyjwt[crypto]==2.4.0 ; python_version >= "3.7" and python_version < "4.0"
pyparsing==2.4.7 ; python_version >= "3.7" and python_version < "4.0"
pysocks==1.7.1 ; python_version >= "3.7" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.7" and python_version < "4.0"
pywin32==303 ; python_version >= "3.7" and platform_system == "Windows" and python_version < "4.0"
remote-pdb==2.1.0 ; python_version >= "3.7" and python_version < "4.0"
requests-oauthlib==1.3.0 ; python_version >= "3.7" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.7" and python_version < "4"
six==1.16.0 ; python_version >= "3.7" and python_version < "4.0"
solnlib==4.8.2 ; python_version >= "3.7" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.7" and python_version < "4.0"
splunk-sdk==1.6.18 ; python_version >= "3.7" and python_version < "4.0"
splunktalib==3.0.3 ; python_version >= "3.7" and python_version < "4.0"
splunktaucclib==6.0.6 ; python_version >= "3.7" and python_version < "4.0"
typing-extensions==4.0.1 ; python_version >= "3.7" and python_version < "4.0"
u-msgpack-python==2.7.0 ; python_version >= "3.7" and python_version < "4.0"
uamqp==1.6.0 ; python_version >= "3.7" and python_version < "4.0"
urllib3==1.26.9 ; python_version >= "3.7" and python_version < "4"
